more work on pathmaker

+++ app/android/SFUMaps/app/src/main/java/me/gurinderhans/sfumaps/utils/MapTools.java
@@ -218,7 +218,7 @@ public class MapTools {
 	/**
 	 * @return - distance in meters
 	 */
-	public static float LatLngDistance(double lat1, double lng1, double lat2, double lng2) {
+	public static double LatLngDistance(double lat1, double lng1, double lat2, double lng2) {
 		double earthRadius = 6371000; //meters
 		double dLat = Math.toRadians(lat2 - lat1);
 		double dLng = Math.toRadians(lng2 - lng1);
@@ -226,7 +226,7 @@ public class MapTools {
 				Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) *
 						Math.sin(dLng / 2) * Math.sin(dLng / 2);
 		double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-		return (float) (earthRadius * c);
+		return earthRadius * c;
 	}
